backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1
supabase==2.5.0
anthropic==0.28.0
razorpay==1.4.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
pypdf==4.2.0
Pillow==10.3.0
pytesseract==0.3.10
python-dotenv==1.0.1
